pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "swallow"
version = "0.1.0"
description = "CLI for a stateful AI workflow system with orchestration, harness runtime, and persistent task state."
readme = "README.md"
requires-python = ">=3.11"
license = { text = "TBD" }
authors = [{ name = "Rocio" }]
dependencies = [
    "httpx>=0.27,<1",
]

[project.optional-dependencies]
dev = [
    "pytest-asyncio>=0.23,<1",
]
vec = [
    "sqlite-vec>=0.1.9,<0.2",
]

[project.scripts]
swl = "swallow.surface_tools.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
"swallow.provider_router" = ["route_policy.default.json", "routes.default.json"]

[tool.pytest.ini_options]
markers = [
    "eval: quality evaluation tests (deselected by default)",
]
addopts = "-m 'not eval'"
